Add data validation, data annotation

+++ api/Controllers/CommentController.cs
@@ -24,6 +24,11 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -34,6 +39,11 @@ namespace api.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+             if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if (comment == null)
@@ -47,9 +57,14 @@ namespace api.Controllers
         [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+             if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             if(!await _stockRepo.StockExists(stockId))
             {
-                return BadRequest("Stock does not exist");
+                return BadRequest("Stock does not exist!");
             }
 
             var commentModel = commentDto.ToCommentFromCreate(stockId);
@@ -62,11 +77,16 @@ namespace api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+             if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if(comment == null)
             {
-                return NotFound("Comment not found");
+                return NotFound("Comment not found!");
             }
 
             return Ok(comment.ToCommentDto());
@@ -76,11 +96,16 @@ namespace api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+             if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepo.DeleteAsync(id);
 
             if(commentModel == null)
             {
-                return NotFound("Comment does not exist");
+                return NotFound("Comment does not exist!");
             }
 
             return Ok(commentModel);
